src/requirements.txt
fastapi
pydantic
uvicorn
load_dotenv
azure-identity
azure-storage-blob
azure-ai-vision-imageanalysis
azure-core
matplotlib
pillow